Hitting error
Noise is being added from within the file, and so I am using only the first index of the generated supersimp numpy arrays (clean data).

Getting the error - Input type (double) and bias type (float) should be the same. I'm not sure where these types are getting defined from, but i suspect it may be from when the noise is added in the file.

I set up single supersimp to tackle the problem but havent used it yet.