Changed dimensions + final real
I have set up the realistic simulator in readiness to produce patterns of dimension 128x88 instead of 88x128.

I have also started a potential final realistic pattern producer, that produces  patterns based off of the particles impact coordinates on the entire quartz plane and the impact velocity.
This could be useful if we want to create a basic Conv NN to make a guess at the particles velocity/ charge/ impact point and hence energy.