Trying more methods
Custom loss isnt lossing. I think its something to do with how its not able to backpropagate through the function easily and is getting confused.
Made new methods of stating the basic loss function extremely explicitly to test the waters (it worked) and making the def into a class (it didnt).