Making Best Realistic Data Gens
FinalGenerator2 enabled hit_point to be in the x axis now as well. The argument is in form [x,y] where x should be between -0.2 and 0.2 and y should be between 0 and 1.6.
FinalGenerator3, instead of rescaling the 509 z axis data to 100, takes a size 100 window instead. This may not be best but it does not change the ToF info.